# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: English Composition 101 course API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /tutor/{moduleId}/conversation:
    get:
      operationId: getTutorConversation
      tags: [tutor]
      summary: Get or create the tutor conversation for a module
      parameters:
        - name: moduleId
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Conversation with messages
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/TutorConversation"
  /tutor/{moduleId}/message:
    post:
      operationId: sendTutorMessage
      tags: [tutor]
      summary: Send a message to the tutor (SSE stream of assistant response)
      parameters:
        - name: moduleId
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/SendTutorMessageBody"
      responses:
        "200":
          description: SSE stream
          content:
            text/event-stream: {}
  /tutor/{moduleId}/critique:
    post:
      operationId: generateCritiqueAnswer
      tags: [tutor]
      summary: Generate a deliberately weak AI answer for the critique exercise
      parameters:
        - name: moduleId
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: A weak AI-generated answer for the student to critique
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CritiqueAnswer"
components:
  schemas:
...
    TutorMessage:
      type: object
      properties:
        id:
          type: integer
        role:
          type: string
        content:
          type: string
        createdAt:
          type: string
          format: date-time
      required:
        - id
        - role
        - content
        - createdAt
    TutorConversation:
      type: object
      properties:
        id:
          type: integer
        moduleId:
          type: string
        messages:
          type: array
          items:
            $ref: "#/components/schemas/TutorMessage"
      required:
        - id
        - moduleId
        - messages
    SendTutorMessageBody:
      type: object
      properties:
        content:
          type: string
      required:
        - content
    CritiqueAnswer:
      type: object
      properties:
        answer:
          type: string
      required:
        - answer